requirements.txt
asgiref==3.5.2
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.50.0
django-phonenumber-field==6.1.0
gunicorn==20.1.0
oauthlib==3.2.0
phonenumbers==8.12.50
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
